meta: remove script and hmac features
These two features are currently un tested and are not being used in CI

This PR removes the script / hmac features completely.
